+++ python/lsst/ts/mtmount/__init__.py
@@ -30,7 +30,6 @@ from .constants import *
 from .exceptions import *
 from .enums import *
 from .utils import *
-from .limits import *
 from . import field_info
 from . import base_message
 from . import commands