+++ python/lsst/ts/mtmount/mock/__init__.py
@@ -19,6 +19,8 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+from .detailed_settings import *
+from .limits import *
 from .axis_device import *
 from .deployable_device import *
 from .main_power_supply_device import *